refactor: consolidate context(s) in lib
