+++ src/main.rs
@@ -28,38 +28,15 @@
 // OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
 // SOFTWARE.
 
-use std::sync::OnceLock;
-
-use crate::util::resolve_to_absolute_path;
-use anyhow::{bail, Context as AnyhowContext, Result};
-use args::{CliArgs, Command};
-use camino::{Utf8Path, Utf8PathBuf};
+use anyhow::Result;
 use clap::Parser;
-use cooklang::{Converter, CooklangParser, Extensions};
-
-// commands
-mod doctor;
-mod import;
-mod lsp;
-mod pantry;
-mod recipe;
-mod report;
-mod search;
-mod seed;
-mod server;
-mod shopping_list;
 #[cfg(feature = "self-update")]
-mod update;
-
-// other modules
-mod args;
-mod util;
-
-const LOCAL_CONFIG_DIR: &str = "config";
-const APP_NAME: &str = "cook";
-const UTF8_PATH_PANIC: &str = "cook only supports UTF-8 paths.";
-const AUTO_AISLE: &str = "aisle.conf";
-const AUTO_PANTRY: &str = "pantry.conf";
+use cookcli::update;
+use cookcli::{
+    args::{CliArgs, Command},
+    configure_context, configure_logging, doctor, import, lsp, pantry, recipe, report, search,
+    seed, server, shopping_list,
+};
 
 pub fn main() -> Result<()> {
     let args = CliArgs::parse();
@@ -81,102 +58,4 @@ pub fn main() -> Result<()> {
         #[cfg(feature = "self-update")]
         Command::Update(args) => update::run(args),
     }
-}
-
-pub struct Context {
-    base_path: Utf8PathBuf,
-    parser: OnceLock<CooklangParser>,
-}
-
-impl Context {
-    pub fn new(base_path: Utf8PathBuf) -> Self {
-        Self {
-            base_path: base_path,
-            parser: OnceLock::new(),
-        }
-    }
-
-    pub fn parser(&self) -> &CooklangParser {
-        self.parser.get_or_init(configure_parser)
-    }
-
-    pub fn aisle(&self) -> Option<Utf8PathBuf> {
-        let auto = self.base_path.join(LOCAL_CONFIG_DIR).join(AUTO_AISLE);
-
-        tracing::trace!("checking auto aisle file: {auto}");
-
-        auto.is_file().then_some(auto).or_else(|| {
-            let global = global_file_path(AUTO_AISLE).ok()?;
-            tracing::trace!("checking global auto aisle file: {global}");
-            global.is_file().then_some(global)
-        })
-    }
-
-    pub fn pantry(&self) -> Option<Utf8PathBuf> {
-        let auto = self.base_path.join(LOCAL_CONFIG_DIR).join(AUTO_PANTRY);
-
-        tracing::trace!("checking auto pantry file: {auto}");
-
-        auto.is_file().then_some(auto).or_else(|| {
-            let global = global_file_path(AUTO_PANTRY).ok()?;
-            tracing::trace!("checking global auto pantry file: {global}");
-            global.is_file().then_some(global)
-        })
-    }
-
-    pub fn base_path(&self) -> &Utf8PathBuf {
-        &self.base_path
-    }
-}
-
-fn configure_parser() -> CooklangParser {
-    CooklangParser::new(Extensions::empty(), Converter::default())
-}
-
-fn configure_context(args: &CliArgs) -> Result<Context> {
-    let base_path = match args.command {
-        Command::Server(ref server_args) => server_args
-            .get_base_path()
-            .unwrap_or_else(|| Utf8PathBuf::from(".")),
-        Command::ShoppingList(ref shopping_list_args) => shopping_list_args
-            .get_base_path()
-            .unwrap_or_else(|| Utf8PathBuf::from(".")),
-        _ => Utf8PathBuf::from("."),
-    };
-
-    let absolute_base_path = resolve_to_absolute_path(&base_path)?;
-
-    if !absolute_base_path.is_dir() {
-        bail!("Base path is not a directory: {}", absolute_base_path);
-    }
-
-    Ok(Context {
-        base_path: absolute_base_path,
-        parser: OnceLock::new(),
-    })
-}
-
-fn configure_logging(verbosity: u8) {
-    let env_filter = match verbosity {
-        0 => "warn,cook=warn",   // Default: warnings and errors only
-        1 => "info,cook=info",   // -v: info level
-        2 => "debug,cook=debug", // -vv: debug level
-        _ => "trace,cook=trace", // -vvv or more: trace level
-    };
-
-    tracing_subscriber::fmt()
-        .with_env_filter(env_filter)
-        .without_time()
-        .with_target(false)
-        .compact()
-        .with_writer(std::io::stderr)
-        .init();
-}
-
-pub fn global_file_path(name: &str) -> Result<Utf8PathBuf> {
-    let dirs = directories::ProjectDirs::from("", "", APP_NAME)
-        .context("Could not determine home directory path")?;
-    let config = Utf8Path::from_path(dirs.config_dir()).expect(UTF8_PATH_PANIC);
-    let path = config.join(name);
-    Ok(path)
 }
